Move home app over to chemie app
