--- chemie/home/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-
-from .models import FundsApplication
-
-admin.site.register(FundsApplication)
